Automate stub generation and linking
Now, the build pipeline will automatically generate .c and .h files which convert C calls to the OCaml calling conventions. WASM functions are given an internal symbol, which is called by the stub with the correct conventions.

The stub.c file is linked against the generated .o file, and a .h is generated, so it should be as simple as calling the exported functions from a C library now. Hurrah!