Properly mark unallocated registers as clobbered
RBP is but one register that may be clobbered by OCaml. For our purposes, runs of tests are independent. But it may be that in Real Applications, users may expect the contents of their registers to be preserved by function calls ;)

This commit marks all registers not used for arguments or result types as clobbered, so that they are properly saved to / restored from the stack.