Add type hashing function + tests
On a 64-bit machine, can compile up to 20 argument functions.  This is more than
is allowable by stub generation in any case.

It may be that we want to hash functions with larger domains.  This could be
done in the wasm2c style using an RTS. In any case, it's not a priority.

This also adds the infrastructure for OUnit tests, should we want to use them
for anything else.