Rewrite import/export system. (#41)
* Initial work on global imports.

Doesn't compile yet, hence in a branch.

* Codegen for globals

* A *lot* of work completely reworking exports.

Still isn't finished, there's still a fair chunk to do, just need to keep chipping away at it. Tomorrow, hopefully.

* It builds! Time to debug...

* Block test works :) Now for the full run

* Fix some more bugs

(It helps to export the data symbols!!!)

* Frontend cleanup; support custom prefixes

* Pass prefix to cmm_of_wasm, some other fixes

There are still a few bugs to be ironed out. But I'm hopeful the main major re-engineering is basically done now.

* Generate init stub

"init" uses the OCaml calling conventions, which do not guarantee that registers aren't clobbered.

This patch uses the same technique as we do for other cstubs functions to ensure registers are preserved by the init function.

* Ensure memory symbol is only generated for local memory

This was hilarious to debug, since the symbol names were the same! (One was in .data, and one was in .bss).