Work on compiling bonkers names
Unicode support in any programming language is an absolute curse. This is an absolute nightmare to compile, mainly due to Python's fussiness about loading in already-encoded strings.

I've got a reasonable distance but honestly not sure what's going wrong now.

(Been debugging this for hours. It probably shows in my tone.)