Dynamically allocate table contents
We were statically allocating them before (whoops!), which leads to a lot of wasted file space, and this was interacting particularly badly with the spectest module which has the correct layout.

In doing so, we fix the 2 silently-failing tests in elems.wast.

We also expose a segfault in imports.wast, but I'm confident that's to do with #49.