More uniform try/catch using Craise (#52)
Instead of trapping by calling "wasm_rt_trap" directly, we trap by using Craise, which is caught by a top-level handler.

This means we no longer have to propagate trap types; the compiler will figure out all of this for us. It also fixes all the issues with bad join points, and is a lot less brittle.