Make translation environment mutable
A lot of the test applications are taking a while to compile, and I think this is because we're using immutable maps. This is fine for smaller programs, but profiling shows that it can get super inefficient for programs with lots of variables, for example.

This patch makes the stackless pass use a mutable translation environment, favouring arrays over immutable maps.