Smarter SSA (#57)
Fixes #55.

Before, we were propagating *all* local variables, regardless whether or not they were modified, leading to lots of useless bindings.

This patch ensures that only locals that are modified are propagated.

Alas, performance gain on the test suite is modest. But this is a pre-req for constant propagation, which I'm imagining will vastly improve performance.