Constant propagation (#61)
This patch introduces constant propagation. As a result, we can evaluate a fair bit more statically, and also avoid the need to introduce so many local variables.

That said, the compilation performance gain is still modest (we've cut about a minute off running the test suite.) There are a few reasons for this: firstly, it's not possible to evaluate everything statically right now -- for example, without external libraries (which I will investigate), it's not possible to do unsigned comparisons, unsigned division, `ctz`, `popcount`, or `clz`. 

Hopefully DCE will help further.

Treating `f32`s as a lost cause, of course.

* first cut of constant propagation

It doesn't seem to have helped compilation time anywhere near as much as I thought...

* Constant folding of arithmetic expressions

* Specialise some more operations at compile time

Again, modest performance increase...

* Optimise statically-known memory accesses

* Statically resolve relation operations

* Try and evaluate even more statically

This includes div / mod, and all of the unary float64 ops, as well as conditionals.