Default to linear scan allocator.
Even though we're pruning as many redundant let-bindings as possible through the use of constant folding, DCE, and inlining, we still have too many in certain circumstances, which leads to huge compile times using the default OCaml allocator.

In this patch, we default to the linear scan allocator (allowing use of the colouring allocator through a compile flag). This allows the game demo to compile in less than a minute (which is still pretty high, but tractable), reduces `skip-stack-guard-page` to ~8seconds, and shaves a minute off the test suite runtime.

The patch also fixes a bug where the fuel variable is not propagated over an exported call (d'oh!). This in turn fixes #62.

I imagine we can speed things up even more when we end up optimising memory checks.