A lot of work on the Emscripten RTS.
It's getting there, it really is. But it's still trapping on the 5th call to malloc in the atax benchmark -- and I'm not sure why. Same behaviour with dlmalloc and emmalloc.

More debugging to do, but I've been in this office for 9hours straight now on a Saturday, and I really need to eat...