Emscripten RTS (#67)
* Begin working on a minimal Emscripten RTS

Immediate goal with this is to get PolyBenchC compiling, so we can get some numbers.

* Flesh out remaining symbols needed for atax

(untested)

* Fix mistake with compiling globals

The whole point of global initialisers is that so we can mutate the global in the *importing* module -- we were mutating the original global!

* Ensure init() method is exported in header

* A lot of work on the Emscripten RTS.

It's getting there, it really is. But it's still trapping on the 5th call to malloc in the atax benchmark -- and I'm not sure why. Same behaviour with dlmalloc and emmalloc.

More debugging to do, but I've been in this office for 9hours straight now on a Saturday, and I really need to eat...

* Fix memory corruption in Emscripten RTS

PolyBenchC's 2mm benchmark runs now! And I'm hoping most of the others do, too. Hurrah!