Revert SIGFPE opt as its incompatible with DCE.
The testsuite passes again now, even with OCaml's DCE enabled. Hurrah! The issue, however, is that some of the PolyBench benchmarks still get generated with illegal instructions...

(Although, it seems we're on par with the performance of V8 on Atax at least, after the memory optimisation has been enabled ;P)