Use `mmap` for implicit memory bounds checks (#68)
Implements #38 for memory (but not tables).

Also cleans up a lot of spaghetti code for division, which I rewrote when debugging.

This PR enables implicit memory bounds checks by using mmap instead of malloc for linear memory. The key idea is to rely on hardware protection, as described in the original WASM paper, instead of explicit checks (although these can be re-enabled via the --explicit-bounds-checks flag.

Doing so yields significant speedups -- we are on par with (and sometimes beat!) V8 in the benchmarks I've tried.